// src/modules/level/level.enum.ts
export enum LevelEntityEnum {
  Language = 'Language',
  Experience = 'Experience',
  Sector = 'Sector',
}

export enum LevelEnum {
  Beginner = 'Beginner',
  Intermediate = 'Intermediate',
  Advanced = 'Advanced',
  Fluent = 'Fluent',
  MotherTongue = 'Mother tongue',
  Expert = 'Expert',
  Level1 = '<1 year',
  Level2 = '1-2 years',
  Level3 = '2-5 years',
  Level4 = '5-10 years',
  Level5 = '10-20 years',
  Level6 = '>20 years',
}